+++ src/features/shows/features/episode/queries.ts
@@ -3,6 +3,7 @@ import { gql } from '@apollo/client';
 export const QUERY_LIST_UP_NEXT = gql`
   query ListUpNext {
     listUpNext {
+      id
       externalId
       number
       seasonNumber
